+++ app/utils/keybox.js
@@ -8,7 +8,7 @@ import {
 import { hexToByte } from './utils';
 import { derivationPath } from '../config/config';
 
-const getPublicKeyFromSecret = secretKey => getPkFromSeed(hexToByte(secretKey))
+export const getPublicKeyFromSecret = secretKey => getPkFromSeed(hexToByte(secretKey))
   .toString('hex')
   .slice(2)
   .toUpperCase();
@@ -32,19 +32,20 @@ function generateKeyPair(seed, path) {
   };
 }
 
-function seedPhraseToHex(seedPhrase) {
+export function seedPhraseToHex(seedPhrase) {
   return bip39.mnemonicToSeedHex(seedPhrase);
 }
 
-function generateNextKey(seed, index) {
+export function generateNextKey(seed, index) {
   const path = `${derivationPath}${index}'`;
   return {
     name: path,
+    index,
     ...generateKeyPair(seed, path)
   };
 }
 
-function generateKeys(seed, from, to) {
+export function generateKeys(seed, from, to) {
   const keys = [];
   for (let i = from; i < to; i++) {
     keys.push(generateNextKey(seed, i));
@@ -52,7 +53,7 @@ function generateKeys(seed, from, to) {
   return keys;
 }
 
-function initKeys(seed, quantity) {
+export function initKeys(seed, quantity) {
   const keys = [];
   keys.push({
     name: 'Master',
@@ -63,15 +64,20 @@ function initKeys(seed, quantity) {
   return keys;
 }
 
-function generateSeedPhrase() {
-  return bip39.generateMnemonic();
+export function findKeyIndex(seed, publicKey) {
+  const step = 10;
+  const steps = 10;
+  let key;
+  for (let i = 0; i < steps; ++i) {
+    key = generateKeys(seed, step * i, step * (i + 1))
+      .find(k => k.publicKey === publicKey);
+    if (key) {
+      return key.index;
+    }
+  }
+  return -1;
 }
 
-export {
-  getPublicKeyFromSecret,
-  seedPhraseToHex,
-  initKeys,
-  generateNextKey,
-  generateSeedPhrase,
-  generateKeys
-};
+export function generateSeedPhrase() {
+  return bip39.generateMnemonic();
+}