+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+# [0.3.0] - 2019-01-30
+### Added
+- Firefox extension support
+- Transaction fee info
+- Account balance in selector and transfer form
+- Automatic key generation during account import
+### Changed
+- Password min length info on restore page
+- Build process
+### Removed
+- Password validation during adding items
+
 # [0.2.1] - 2019-01-16
 ### Added
 - Account menu chevron
@@ -44,9 +56,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Storage encryption
 - Docs
 
-[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.3.0...HEAD
 
-[0.2.0]: https://github.com/adshares/ads-browser-wallet/compare/v0.2.0...v0.2.1
+[0.3.0]: https://github.com/adshares/ads-browser-wallet/compare/v0.2.1...v0.3.0
+[0.2.1]: https://github.com/adshares/ads-browser-wallet/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/adshares/ads-browser-wallet/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/adshares/ads-browser-wallet/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/adshares/ads-browser-wallet/releases/tag/v0.1.0